Fix to compiler warnings

+++ src/unit/unit_manager.cpp
@@ -183,7 +183,7 @@ void CUnitManager::Save(CFile &file) const
 	for (std::vector<CUnit *>::const_iterator it = units.begin(); it != units.end(); ++it) {
 		const CUnit &unit = **it;
 		if (unit.Container && unit.Container->Character && unit.Container->Character->Persistent && unit.Container->HasInventory() && unit.Type->BoolFlag[ITEM_INDEX].value) { // don't save items for persistent heroes
-			file.printf("{Slot = %d, FreeCycle = %u}, ", UnitNumber(unit), GameCycle);
+			file.printf("{Slot = %d, FreeCycle = %u}, ", UnitNumber(unit), (unsigned int) GameCycle);
 		}
 	}
 	//Wyrmgus end
